requirements.txt
langchain
langchain-core
langchain_openAI
python-dotenv